Importing Dashboard and Adjustments
Importing the dashboard from the automotive application. Changed the color scheme to red and white. Imported the elements from the skeleton to the code file, so there's like 300 lines that do absolutely nothing but save the program from errors, but 30+ warnings. The welcome portion works on login and it should on signup but I didn't check. Password regex now works properly. Started adding the code that will operate the first table. Created a complete medicine class to hold each medicine's information.