Lots of changes, icons, maximize button, functioning table
Adding font awesome logos. went through the code and updated the sql and now update item, delete item, and add item works. Fixed page status to be 0, 1, 2, instead of 0, 1, 3. Simplified some code in checkValidUser and removed some souts in resetPassword. Changed the choiceBox initial values to be shorter so they are in frame. Some basic formatting. Fixed the window dragging problem. Adding a maximize/normalize button like in windows where pressing it goes fullscreen and pressing it again goes back to the normal window. It works but when maximized the top bar doesn't reach all the way across and the content isn't centered.