Doin big things
Added several photos for medicines and logos. Added dates whenever an item is updated or added and updated sql accordingly, added image choosing and image getting when an item is clicked on the table, added imagePath to the medicine class and added a create image method to the utils class, simplified the window normalize and maximize methods to just one that does everything by itself, created a get medicine method that simplifies the process of creating a list of data that is repeated twice.