Lots of changes and formatting
Changing all choice boxes to combo boxes for consistency, formatting code on all pages to fit better and be shorter and making all the alert box references be on multiple lines to look better and be easier to read, any methods that use a lot of parameters go onto multiple lines to make it easier to read. Added a format double method in Utils to make all doubles have #.## format. Removed redundant assigning null to null values in user class, any long sql lines in SQLUtils I put on multiple lines to make it easier to read and look like actual sql. Shortening array initializations. Filter table works better and I was working on clear table, clear cart, and remove from cart but realized it's not ever required in the instructions so I stopped.